backend/requirements.txt
fastapi[standard]>=0.115.0
sqlmodel>=0.0.22
alembic>=1.14.0
asyncpg>=0.30.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
PyJWT[crypto]>=2.8.0
Jinja2>=3.1.0
openai>=1.60.0
sse-starlette>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
